backend/requirements.txt
# FastAPI & Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# API Clients & LLM
cohere>=5.0.0
huggingface-hub>=0.17.0

# AI/ML & Agents
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Vector Database
qdrant-client>=1.7.0,<2.0

# NLP & Text Processing
beautifulsoup4>=4.12.0
nltk>=3.8.0
tiktoken>=0.5.0

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
email-validator>=2.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
aiohttp>=3.8.0
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.12.0
ruff>=0.1.0
